Fix passing of UINT_MAX and ULONG_MAX from C++ to JS function. LLVM doesn't have u32 type, so UINT_MAX literal comes out as 'i32 -1' literal to JS, which whould treat it as a negative -1.0 double.

+++ system/include/emscripten/bind.h
@@ -40,8 +40,8 @@ namespace emscripten {
             void _embind_register_integer(
                 TYPEID integerType,
                 const char* name,
-                int minRange,
-                int maxRange);
+                long minRange,
+                unsigned long maxRange);
 
             void _embind_register_float(
                 TYPEID floatType,
